% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRxiv-package.R
\docType{package}
\name{aRxiv-package}
\alias{aRxiv}
\alias{aRxiv-package}
\title{aRxiv: Interface to the arXiv API}
\description{
An interface to the API for 'arXiv', a repository of electronic preprints for computer science, mathematics, physics, quantitative biology, quantitative finance, and statistics.
}
\section{vignette}{

\href{https://docs.ropensci.org/aRxiv/articles/aRxiv.html}{aRxiv tutorial}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/aRxiv/}
  \item \url{https://github.com/ropensci/aRxiv}
  \item Report bugs at \url{https://github.com/ropensci/aRxiv/issues}
}

}
\author{
\strong{Maintainer}: Karl Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
  \item Karthik Ram \email{karthik.ram@gmail.com} (\href{https://orcid.org/0000-0002-0233-1757}{ORCID})
}

}
\keyword{internal}
