% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_cats-data.R
\docType{data}
\name{arxiv_cats}
\alias{arxiv_cats}
\title{arXiv subject classifications}
\format{
A data frame with five columns: the abbreviations of the
subject classifications (\code{category}), the field of study,
subfield of study (within Physics; \code{NA} otherwise), a short
description, and a longer description.
}
\source{
\url{https://arxiv.org/category_taxonomy}
}
\usage{
data(arxiv_cats)
}
\description{
arXiv subject classifications: their abbreviations and
corresponding descriptions.
}
\examples{
arxiv_cats
}
\keyword{datasets}
