% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arxiv_open}
\alias{arxiv_open}
\title{Open abstract for results of arXiv search}
\usage{
arxiv_open(search_results)
}
\arguments{
\item{search_results}{Data frame of search results, as returned from \code{\link{arxiv_search}}}
}
\value{
(Invisibly) Vector strings with URLs of abstracts opened.
}
\description{
Open, in web browser, the abstract pages for each of set of arXiv search results.
}
\details{
The R option \code{"arxiv_max2open"} defines the maximum
number of pages to open; if missing, the default is 5.
\bold{Don't try to open too many at once!}
}
\examples{
z <- arxiv_search("au:Broman AND cat:stat.AP")
\donttest{arxiv_open(z)}
}

