% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_arxiv_connect.R
\name{can_arxiv_connect}
\alias{can_arxiv_connect}
\title{Check for connection to arXiv API}
\usage{
can_arxiv_connect(max_time = 5)
}
\arguments{
\item{max_time}{Maximum wait time in seconds}
}
\value{
Returns TRUE if connection is established and FALSE
otherwise.
}
\description{
Check for connection to arXiv API
}
\examples{
\donttest{
if(interactive()) {
    can_arxiv_connect(2)
} }

}
