% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{create_filename}
\alias{create_filename}
\title{Create filename from arxiv URL}
\usage{
create_filename(arx_url)
}
\arguments{
\item{arx_url}{a string containing the URL for arxiv paper}
}
\value{
A filename containing the calibre style title for arxiv URL
}
\description{
Arxiv holds a collection of pre-prints for a variety of scientitic
topics and research areas. Calibre is a handy tool to save and
catalog different articles and books in a single location.
This function downloads metadata from an arxiv URL and creates a file
name corresponding to the paper in the Calibre standard format.
}
