% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{download_paper}
\alias{download_paper}
\title{Dowload paper locally}
\usage{
download_paper(arx_url, loc)
}
\arguments{
\item{arx_url}{a string containing the arxiv URL}

\item{loc}{a string containing the file path to save it to}
}
\description{
Downloads paper and renames it using the \code{create_filename} function
}
