% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{query_prefixes}
\alias{query_prefixes}
\title{arXiv query field prefixes}
\format{A data frame with two columns: the \code{prefix} and corresponding
\code{explanation}.}
\usage{
data(geneExpr)
}
\description{
Possible prefixes that correspond to different fields in arXiv searches.
}
\examples{
data(query_prefixes)
}
\author{
Karl W Broman
}
\keyword{datasets}

