% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{query_terms}
\alias{query_terms}
\title{arXiv query field terms}
\format{A data frame with two columns: the \code{term} and corresponding
\code{description}.}
\source{
\url{http://arxiv.org/help/api/user-manual}
}
\usage{
data(query_terms)
}
\description{
Possible terms that correspond to different fields in arXiv searches.
}
\examples{
query_terms
}
\author{
Karl W Broman
}
\keyword{datasets}

