% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{arxiv_cats}
\alias{arxiv_cats}
\title{arXiv subject classifications}
\format{A data frame with two columns: the abbreviations of the
subject classifications (\code{abbreviation}) and the corresponding
description (\code{description}).}
\source{
\url{http://arxiv.org/help/api/user-manual}
}
\usage{
data(arxiv_cats)
}
\description{
arXiv subject classifications: their abbreviations and
corresponding descriptions.
}
\examples{
arxiv_cats
}
\keyword{datasets}

