% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arxiv_count}
\alias{arxiv_count}
\title{Count number of results for a given search}
\usage{
arxiv_count(query = NULL, id_list = NULL)
}
\arguments{
\item{query}{Search pattern as a string}

\item{id_list}{List of arXiv doc IDs, as comma-delimited string}
}
\value{
Number of results (integer)
}
\description{
Count the number of results for a given search. Useful to check
before attempting to pull down a very large number of records.
}
\examples{
# count papers in category stat.AP (applied statistics)
\dontshow{options(aRxiv_delay=1)}
arxiv_count(query = "cat:stat.AP")
arxiv_count(query = 'au:"Peter Hall"')
}

