% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arxiv_open}
\alias{arxiv_open}
\title{Open abstract for results of arXiv search}
\usage{
arxiv_open(search_results)
}
\arguments{
\item{search_results}{Data frame of search results, as returned from \code{\link{arxiv_search}}}
}
\value{
(Invisibly) Vector of character strings with URLs of
abstracts opened.
}
\description{
Open, in web browser, the abstract pages for each of set of arXiv search results.
}
\details{
The R option \code{"arxiv_max2open"} defines the maximum
number of pages to open; if missing, the default is 20.

There is also a delay between calls to
\code{\link[utils]{browseURL}}, with the amount taken from the R
option \code{"arxiv_delay"} (in seconds); if missing, the default
is 3 sec.
}
\examples{
z <- arxiv_search('au:"Peter Hall" AND ti:deconvolution')
\donttest{arxiv_open(z)}
}

