% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arxiv_search}
\alias{arxiv_search}
\title{The main search function for aRxiv}
\usage{
arxiv_search(query = NULL, id_list = NULL, start = NULL, end = NULL,
  sort_by = c("relevance", "lastUpdatedDate", "submittedDate"),
  ascending = TRUE)
}
\arguments{
\item{query}{Search pattern; submit multiple terms as a concactenated string.}

\item{id_list}{List of arXiv doc IDs, as comma-delimited string}

\item{start}{An offset for the start of search}

\item{end}{Where to end search results}

\item{sort_by}{How to sort the results}

\item{ascending}{If TRUE, sort in increasing order; else decreasing}
}
\value{
Parse XML result as a list
}
\description{
Allows for progammatic searching of the arXiv pre-print repository.
}
\examples{
z <- arxiv_search(query = "au:Broman AND cat:stat.AP", start=0, end=10)
z$totalResults
titles <- sapply(z[names(z)=="entry"], function(a) a$title)
}

