% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arxiv_search}
\alias{arxiv_search}
\title{The main search function for aRxiv}
\usage{
arxiv_search(query = NULL, id_list = NULL, start = 0, end = 10,
  sort_by = c("relevance", "lastUpdatedDate", "submittedDate"),
  ascending = TRUE, batchsize = 500, force = FALSE)
}
\arguments{
\item{query}{Search pattern as a string}

\item{id_list}{List of arXiv doc IDs, as comma-delimited string}

\item{start}{An offset for the start of search}

\item{end}{Where to end search results}

\item{sort_by}{How to sort the results}

\item{ascending}{If TRUE, sort in increasing order; else decreasing}

\item{batchsize}{Maximum number of records to request at one time}

\item{delay}{Time between requests, in seconds}

\item{force}{If TRUE, force search request even if it seems extreme}
}
\value{
Parse XML result as a list
}
\description{
Allows for progammatic searching of the arXiv pre-print repository.
}
\examples{
# search for author Broman and category stat.AP (applied statistics)
z <- arxiv_search(query = "au:Broman AND cat:stat.AP", start=0, end=10)
z$totalResults
sapply(z[names(z)=="entry"], function(a) a$title)

# search for a set of documents by arxiv identifiers
z <- arxiv_search(id_list = "1403.3048,1402.2633,1309.1192")
# DOI if available
sapply(z[names(z)=="entry"], function(a) a$doi)
}

